+++ src/app/detail/[task_id]/[user_type]/actions.ts
@@ -108,23 +108,10 @@ export const deleteComment = async (token: string, id: string) => {
   revalidateTag('getActivities')
 }
 
-export const getSignedUrlFile = async (token: string, filePath: string) => {
-  const res = await fetch(`${apiUrl}/api/attachments/sign-url?token=${token}&filePath=${filePath}`)
-  const data = await res.json()
-  return data.signedUrl
-}
-
 export const postScrapImage = async (token: string, payload: ScrapImageRequest) => {
   console.log('posting')
   await fetch(`${apiUrl}/api/tasks/${payload.taskId}/scrap-image/?token=${token}`, {
     method: 'POST',
     body: JSON.stringify(payload),
   })
-}
-
-async function getSignedUrlUpload(token: string, fileName: string) {
-  const res = await fetch(`${apiUrl}/api/attachments/upload?token=${token}&fileName=${fileName}`)
-
-  const data = await res.json()
-  return data.signedUrl
 }
