+++ src/app/detail/ui/TaskEditor.tsx
@@ -18,8 +18,7 @@ import { SupabaseActions } from '@/utils/SupabaseActions'
 import { generateRandomString } from '@/utils/generateRandomString'
 import { TaskResponse } from '@/types/dto/tasks.dto'
 import { ScrapImageRequest } from '@/types/common'
-import { getSignedUrlFile } from '@/app/detail/[task_id]/[user_type]/actions'
-import { getFilePathFromUrl } from '@/utils/signedUrlReplacer'
+
 import { getSignedUrlUpload } from '@/app/actions'
 import { deleteEditorAttachmentsHandler, uploadImageHandler } from '@/utils/inlineImage'
