+++ src/utils/inlineImage.ts
@@ -3,11 +3,12 @@ import React from 'react'
 import { generateRandomString } from '@/utils/generateRandomString'
 import { ISignedUrlUpload } from '@/types/interfaces'
 import { getSignedUrlUpload } from '@/app/actions'
-import { getSignedUrlFile, postScrapImage } from '@/app/detail/[task_id]/[user_type]/actions'
+import { postScrapImage } from '@/app/detail/[task_id]/[user_type]/actions'
 import { ScrapImageRequest } from '@/types/common'
 import { getFilePathFromUrl } from '@/utils/signedUrlReplacer'
+import { getSignedUrlFile } from '@/app/page'
 
-export const uploadImageHandler = async (file: File, token: string, task_id?: string): Promise<string> => {
+export const uploadImageHandler = async (file: File, token: string, task_id: string | null): Promise<string> => {
   const supabaseActions = new SupabaseActions()
   const fileName = generateRandomString(file.name)
 
@@ -18,12 +19,12 @@ export const uploadImageHandler = async (file: File, token: string, task_id?: st
   return url
 }
 
-export const deleteEditorAttachmentsHandler = async (url: string, token: string, task_id?: string) => {
+export const deleteEditorAttachmentsHandler = async (url: string, token: string, task_id: string | null) => {
   const filePath = await getFilePathFromUrl(url)
   if (filePath) {
     const payload: ScrapImageRequest = {
       filePath: filePath,
-      taskId: task_id ? task_id : null,
+      taskId: task_id,
     }
     postScrapImage(token, payload)
   }
